add bottom navbar, add _ to cart state/cubit (separate from the main cart cubit), ux improvements

+++ lib/repositories/cart_repositories.dart
@@ -18,8 +18,8 @@ class CartRepositories {
       final foodObj = Food(
         title: foodItem['title'],
         description: foodItem['description'],
-        price: foodItem['price'],
-        originalPrice: foodItem['original_price'],
+        price: foodItem['price'].toDouble(),
+        originalPrice: foodItem['original_price'].toDouble(),
         imageName: foodItem['image_name'],
         id: foodItem['id'],
       );
